router 설정 수정
